Lots of game logic

+++ app/src/main/java/com/chumaribelle/quantumcargame/SuperpositionCar.java
@@ -7,11 +7,11 @@ import android.graphics.RectF;
 
 public class SuperpositionCar extends CarSprite {
 
-    int pos;
+    float pos;
     boolean inBounds;
     float top, bottom;
 
-    public SuperpositionCar(CarSprite car, float dX, Bitmap bitmap, int pos) {
+    public SuperpositionCar(CarSprite car, float dX, Bitmap bitmap, float pos) {
         super((int) car.left, (int) car.top, (int) car.right, (int) car.bottom, dX, 0, bitmap);
         top = car.top;
         bottom = car.bottom;
